Extract editor export workflow hook
